getweather/requirements.txt
openmeteo-requests
requests-cache
retry-requests
numpy
pandas
python-dotenv
